+++ src/ckernel/render_context/opengl_renderer/context.py
@@ -1,9 +1,9 @@
 from ckernel.render_context._renderer import Renderer
 import ckernel.render_context.opengl_renderer.opengl_hooker as hooked_opengl
-from ckernel.glfw_context.none_context import NoneContextManager
+from .context_stack import OpenglContextStack
 
 
-class OpenglRenderer(Renderer):
+class OpenglContext(Renderer):
     def __init__(self, context):
         self._cntxt_manager = context
 
@@ -13,7 +13,7 @@ class OpenglRenderer(Renderer):
 
         :return:
         """
-        OpenglContextStack._put_current(self)
+        OpenglContextStack.put_current(self)
         with self._cntxt_manager.glfw as glfw:
             glfw.make_context_current()
         return hooked_opengl
@@ -27,49 +27,8 @@ class OpenglRenderer(Renderer):
         :param exc_tb:
         :return:
         """
-        OpenglContextStack._pop_current()
+        OpenglContextStack.pop_current()
         # return binding
-        with OpenglContextStack._get_current()._cntxt_manager.glfw as glfw:
+        with OpenglContextStack.get_current()._cntxt_manager.glfw as glfw:
             glfw.make_context_current()
 
-
-class OpenglNoneContext(OpenglRenderer):
-
-    def __init__(self):
-        self._cntxt_manager = NoneContextManager()
-
-
-# dont know context binding is a common thing so placing class for opengl
-class OpenglContextStack:
-    # reserve None context as first element
-    __stack = [OpenglNoneContext()]
-
-    @classmethod
-    def _get_current(cls):
-        """
-        return current context
-
-        :return:
-        """
-        return cls.__stack[-1]
-
-    @classmethod
-    def _put_current(cls, context):
-        """
-        add current to the stack
-
-        :param context:
-        :return:
-        """
-        cls.__stack.append(context)
-
-    @classmethod
-    def _pop_current(cls):
-        """
-        remove top context
-
-        To return context to idle, None context is never removed
-        :return:
-        """
-        if 1 < len(cls.__stack):
-            cls.__stack.pop()
