+++ src/ckernel/render_context/opengl_renderer/none_context.py
@@ -1,0 +1,9 @@
+from ckernel.glfw_context.none_context import NoneContextManager
+
+
+class OpenglNoneContext:
+
+    def __init__(self):
+        self._cntxt_manager = NoneContextManager()
+
+
