--- src/ckernel/render_context/renderers.py
@@ -1,1 +1,0 @@
-from .opengl_renderer.context import OpenglRenderer
