+++ src/doodler/basics.py
@@ -1,4 +1,5 @@
-import wkernel.hooked.openglHooked as gl
+import OpenGL.GL as gl
+
 import wkernel.pipeline.nodes as node
 from global_tools import Singleton
 from wkernel.env.components.bits import DrawInterface
