+++ src/wkernel/__init__.py
@@ -4,11 +4,3 @@ Universal Visualization Tool(working title)
 
 from wkernel.env.components.bits import DrawInterface
 from .env.window import Window
-from .hooked import glfwHooked as glfw
-from .hooked import openglHooked as gl
-#
-# class Window(Windows):
-#     """
-#     Front face of 'components' class.
-#     """
-#     pass
