--- src/wkernel/hooked/__init__.py
@@ -1,2 +1,0 @@
-import wkernel.hooked.glfwHooked as glfw
-import wkernel.hooked.openglHooked as gl
