+++ src/wkernel/pipeline/nodes/opengl/_opengl.py
@@ -1,5 +1,4 @@
 from JINTFP import *
-from wkernel.hooked import *
 from ...data_types import *
