+++ src/wkernel/pipeline/nodes/opengl/primitive/program_shader_nodes.py
@@ -1,4 +1,6 @@
 from .._opengl import *
+import OpenGL.GL as gl
+
 
 class ConPrgrm(OpenglNode):
     """
