+++ src/wkernel/pipeline/nodes/opengl/primitive/render_nodes.py
@@ -1,4 +1,5 @@
 from .._opengl import *
+import OpenGL.GL as gl
 
 
 class RenderComponent(OpenglNode):
